refactor: add KVPbCrudApi::crud_try_insert() and crud_try_upsert()
